requirements.txt
# Core dependencies
python-dotenv
pydantic>=2.6.1,<3.0.0

# MCP dependencies
fastmcp>=0.1.0
mcp>=0.1.0

# API Server dependencies  
fastapi
uvicorn

# RAG dependencies
langchain>=0.1.0
langchain-community>=0.0.10
qdrant-client>=1.7.0
pypdf2>=3.0.0
sentence-transformers>=2.2.0
openai>=1.0.0

# Agent framework
crewai>=0.30.0
crewai-tools>=0.4.0

# External APIs
requests>=2.31.0

# Utils
numpy>=1.24.0
pandas>=2.0.0
tiktoken>=0.5.0
nest-asyncio>=1.5.0

# Frontend
streamlit>=1.29.0